+++ 00-Unidades/04_while/Ejercicios_While/While_app_07.py
@@ -43,7 +43,7 @@ class App(customtkinter.CTk):
             
             if numero == None:
                 break
-            numero=int(numero)
+            numero=float(numero)
             acumulador+=numero
             contador+=1
